update log messages
